refactor(cli): move project commands to capability
